% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{appropriateness}
\alias{appropriateness}
\title{Appropriateness of a statement}
\usage{
appropriateness(.data, qualifier, statement)
}
\arguments{
\item{.data}{\code{fuzzy_df}\cr
Fuzzy table to work on.}

\item{qualifier}{\code{fuzzy statement}\cr
Grouping of records towards which to compute appropriateness.}

\item{statement}{\code{fuzzy statement}\cr
A statement to evaluate.}
}
\value{
A single numeric value between -1 and 1, the larger the more
appropriate the statement is.
}
\description{
TODO
}
\examples{
# Are obese people younger?
PimaIndiansDiabetes2 \%>\%
  define_quality(age = qual_age(), mass = qual_bmi()) \%>\%
  appropriateness(bmi \%is\% obese, age \%is\% young)

}
