% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{define_quality}
\alias{define_quality}
\title{Define meanings of columns}
\usage{
define_quality(.data, ...)
}
\arguments{
\item{.data}{\code{data.frame}\cr
Table (possibly fuzzy) to define column meanings for.}

\item{...}{\code{fuzzy_quality(1)}\cr
Fuzzy qualities to define, each for one column. Must be named after target
columns.}
}
\value{
A \code{fuzzy_df} with qualities defined for specified columns (possibly
more, if \code{.data} was a \code{fuzzy_df} with some already defined qualities).
}
\description{
TODO
}
\examples{
# The key building block of fuzzydf package
infert \%>\%
  define_quality(age = qual_age())

}
