% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifier.R
\name{fuzzy-modifier}
\alias{fuzzy-modifier}
\alias{mod_very}
\alias{mod_somewhat}
\title{Apply a modifier to a quality}
\usage{
mod_very(.col)

mod_somewhat(.col)
}
\arguments{
\item{.col}{\code{fuzzy_quality(1)}\cr
Fuzzy quality to modify.}
}
\value{
Fuzzy values of a column, but appropriately modified.
}
\description{
TODO
}
\examples{
data(PimaIndiansDiabetes2, package = "mlbench")
PimaIndiansDiabetes2 \%>\%
  define_quality(age = qual_age(), mass = qual_bmi()) \%>\%
  truth_of(q_few(bmi \%is\% mod_very(obese)))

}
