% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{truth_of}
\alias{truth_of}
\title{Evaluate truth of a statement}
\usage{
truth_of(.data, proposition)
}
\arguments{
\item{.data}{\code{fuzzy_df}\cr
Fuzzy table to work on.}

\item{proposition}{\code{fuzzy quantifier}\cr
Statement about data to evaluate.}
}
\value{
A single numeric value between 0 and 1, the higher the more true the
statement is.
}
\description{
TODO
}
\examples{
# Is it true that most Star Wars characters are old?
starwars \%>\%
  define_quality(birth_year = qual_age()) \%>\%
  truth_of(q_most(age \%is\% old))

# It's possible to evaluate complex statements:
data(PimaIndiansDiabetes2, package = "mlbench")
PimaIndiansDiabetes2 \%>\%
  define_quality(age = qual_age(), mass = qual_bmi()) \%>\%
  truth_of(q_most(age \%is\% young | bmi \%is\% obese))

}
